Style Fix
put the style outside the component
