+++ renta_facil_colima/src/View/components/SplashScreen.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import {View, Image, StyleSheet} from 'react-native';
+import {View, } from 'react-native';
 import LogoRFC from './LogoRFC';
+import styles from '../../styles/SplashScreenStyle';
 
 const SplashScreen = () => {
   return (
@@ -12,26 +13,5 @@ const SplashScreen = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FFFFFF',
-  },
-  logo: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    position: 'absolute',
-    top: 0,
-    bottom: 135,
-  },
-  image: {
-    width: '100%',
-    height: '100%',
-    resizeMode: 'contain',
-  },
-});
 
 export default SplashScreen;
