+++ renta_facil_colima/src/screens/MainScreen.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import {StyleSheet, Text, View} from 'react-native';
 import LogoRFC from '../View/components/LogoRFC';
 import BasicButton from '../View/buttons/BasicButton';
+import styles from '../styles/MainScreenStyle';
 
 export const MainScreen = () => {
   return (
@@ -22,32 +23,3 @@ export const MainScreen = () => {
   );
 };
 
-const styles = StyleSheet.create({
- container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FFFFFF',
-  },
-  logo: {
-    flex: 0,
-    justifyContent: 'center',
-    alignItems: 'center',
-    position: 'absolute',
-    top: 10,
-    bottom: 20,
-  },
-  AppName: {
-    textAlign: 'center',
-  },
-  Botones: {
-    position: 'absolute', // Posición absoluta en relación con el contenedor principal
-    bottom: 0, // Se coloca en la parte inferior
-    left: 0, // Se ajusta a la izquierda
-    right: 0, // Se ajusta a la derecha
-    flexDirection: 'row', // Los componentes se colocarán en una fila
-    justifyContent: 'space-between', // Espacio entre los componentes
-    paddingHorizontal: 30, // Espacio horizontal
-    paddingBottom: 60,
-  },
-});
